moving old attribute cache to be called internal::MapCache
