+++ src/wmtk/attribute/Accessor.hpp
@@ -30,7 +30,7 @@ public:
     friend class wmtk::PointMesh;
     using Scalar = T;
 
-    friend class AttributeCache<T>;
+    friend class internal::AttributeMapCache<T>;
     using BaseType = AccessorBase<T, Dim>;
     using CachingBaseType = CachingAccessor<T, Dim>;
